// cli/fs/src/interfaces.ts
export enum ENodeType {
	FILE = 'file',
	DIR = 'dir',
	SYMLINK = 'symlink',
	OTHER = 'other',
	BLOCK = 'block'
}

/**
 * Native errors.
 * @todo : replace with errno.
 */
export let EError: any = {
	NONE: 'None',
	EXISTS: 'EEXIST',
	PERMISSION: 'EACCES',
	NOEXISTS: 'ENOENT',
	CROSS_DEVICE: 'EXDEV'
};

/////////////////////////////////////////////////////////
//
//  Data structures
//
export interface INode {
	name: string
	type: ENodeType | string
	size: number
	accessTime?: Date
	modifyTime?: Date
	changeTime?: Date
	birthTime?: Date
	absolutePath?: string
	mode?: number
	pointsAt?: string
	relativePath?: string
	children: INode[]
	total?: number
	checksum?: string
	mime?: string
}
/**
 * The options for "inspect".
 *
 * @export
 * @interface IInspectOptions
 */
export interface IInspectOptions {
	checksum?: string
	mode?: boolean
	times?: boolean
	absolutePath?: boolean
	symlinks?: boolean | string
	size?: boolean
	mime?: boolean
}